requirements.txt
pytest
transformers 
faiss-cpu
flask 
torch
